init: PluginFactory support
Some plugins operate by wrapping one or more separately-maintained plugins,
and do so by instantiating them directly. This can be problematic because
the pipeline's execution context is not propagated to the inner plugin, which
can cause features that rely on execution context such as pipeline settings
to not work as expected.

This support adapter introduces a PluginFactory, that can be used to
instantiate inner plugins that are pre-contextualized with the outer plugin's
execution context.

By including this support adapter into a plugin, we ensure that the plugin
has a `LogStash::Plugin#plugin_factory`, which can be used to load
contextualized proxies for plugin classes, which in turn can be instantiated
in the normal way, replacing the direct instantiation of the plugin class:

~~~ diff
   def register
     grok_args = {}

-    @inner = ::LogStash::Filters::Grok.new(grok_args)
+    @inner = plugin_factory.filter('grok').new(grok_args)
  end
~~~